uniform the type format

+++ src/os/newstore/NewStore.cc
@@ -1420,7 +1420,7 @@ int NewStore::fiemap(
       continue;
     }
 
-    unsigned x_len = len;
+    uint64_t x_len = len;
     if (op != oend &&
 	op->first > offset &&
 	op->first - offset < x_len) {
